move common logic into test helpers
Test-tag: hw,medium,dmg,control,engine_auto_restart
