+++ src/tests/ftest/control/engine_auto_restart_advanced.py
@@ -18,35 +18,6 @@ class EngineAutoRestartAdvanced(ControlTestBase):
     :avocado: recursive
     """
 
-    def setUp(self):
-        """Set up each test case."""
-        super().setUp()
-        self.dmg = self.get_dmg_command()
-
-    def get_all_ranks(self):
-        """Get list of all ranks in the system."""
-        return list(self.server_managers[0].ranks.keys())
-
-    def get_rank_state(self, rank):
-        """Get the state of a rank.
-
-        Args:
-            rank (int): Rank number
-
-        Returns:
-            str: Current state of the rank
-        """
-        data = self.dmg.system_query(ranks=f"{rank}")
-        if data["status"] != 0:
-            self.fail("Cmd dmg system query failed")
-        if "response" in data and "members" in data["response"]:
-            if data["response"]["members"] is None:
-                self.fail("No members returned from dmg system query")
-            for member in data["response"]["members"]:
-                return member["state"].lower()
-        self.fail("No member state returned from dmg system query")
-        return None
-
     def wait_for_rank_state(self, rank, expected_state, timeout=30, check_interval=2):
         """Wait for a rank to reach expected state.
 
@@ -96,7 +67,7 @@ class EngineAutoRestartAdvanced(ControlTestBase):
         :avocado: tags=EngineAutoRestartAdvanced,test_deferred_restart
         """
         # Get configured restart delay from test params
-        restart_delay = 15
+        restart_delay = 20
 
         all_ranks = self.get_all_ranks()
         if len(all_ranks) < 2:
@@ -104,57 +75,34 @@ class EngineAutoRestartAdvanced(ControlTestBase):
 
         test_rank = self.random.choice(all_ranks)
 
-#        restarted, final_state = self.exclude_rank_and_wait_restart(test_rank)
-#
-#        if not restarted:
-#            self.fail("Rank %s did not automatically restart. Final state: %s"
-#                      % (test_rank, final_state))
+        self.log_step("Step 1: Automatic restart of rank %s", test_rank)
 
-        # First exclusion - should restart immediately (no previous restart)
-        self.log_step("Step 1: First exclusion of rank %s", test_rank)
-        self.dmg.system_exclude(ranks=[test_rank], rank_hosts=None)
+        restarted, final_state = self.exclude_rank_and_wait_restart(test_rank)
 
-        # Wait for self-termination
-        if not self.wait_for_rank_state(test_rank, "adminexcluded", timeout=10):
-            self.fail("Rank %s did not self-terminate" % test_rank)
-
-        self.dmg.system_clear_exclude(ranks=[test_rank], rank_hosts=None)
-
-        # Wait for automatic restart
-        self.log_step("Step 2: Waiting for first automatic restart of rank %s", test_rank)
-        if not self.wait_for_rank_state(test_rank, "joined", timeout=30):
-            self.fail("Rank %s did not automatically restart on first exclusion" % test_rank)
-
-#
+        if not restarted:
+            self.fail("Rank %s did not automatically restart. Final state: %s"
+                      % (test_rank, final_state))
 
         first_restart_time = time.time()
         self.log.info("First restart completed at T=%.1f", first_restart_time)
 
         # Second exclusion - should be deferred due to rate-limiting
-        self.log_step("Step 3: Second exclusion of rank %s (should be deferred)", test_rank)
-        self.dmg.system_exclude(ranks=[test_rank], rank_hosts=None)
+        self.log_step("Step 2: Second exclusion of rank %s (should be deferred)", test_rank)
 
-        # Wait for self-termination
-        if not self.wait_for_rank_state(test_rank, "adminexcluded", timeout=10):
-            self.fail("Rank %s did not self-terminate on second exclusion" % test_rank)
+        restarted, final_state = self.exclude_rank_and_wait_restart(test_rank,
+                                                                    expect_restart=False,
+                                                                    timeout=10)
 
-        self.dmg.system_clear_exclude(ranks=[test_rank], rank_hosts=None)
+        if restarted:
+            self.fail("Rank %s unexpectedly restarted. Final state: %s"
+                      % (test_rank, final_state))
 
-        # Verify restart is NOT immediate (should be deferred)
-        self.log_step("Step 4: Verifying restart is deferred (not immediate)")
-        time.sleep(5)  # Wait a bit
-
-        current_state = self.get_rank_state(test_rank)
-        if current_state == "joined":
-            self.fail("Rank %s restarted immediately - rate-limiting not working" % test_rank)
-
-        self.log.info("Confirmed: Restart is deferred (rank still in '%s' state)",
-                      current_state)
+        self.log.info("Confirmed: Restart is deferred (rank still in excluded state)")
 
         # Wait for deferred restart to execute (after delay expires)
         # Add buffer time for processing
-        wait_time = restart_delay + 30  # 10
-        self.log_step("Step 5: Waiting %ss for deferred restart to execute", wait_time)
+        wait_time = restart_delay + 10
+        self.log_step("Step 3: Waiting %ss for deferred restart to execute", wait_time)
 
         if not self.wait_for_rank_state(test_rank, "joined", timeout=wait_time):
             self.fail("Rank %s did not restart after rate-limit delay" % test_rank)
@@ -200,31 +148,31 @@ class EngineAutoRestartAdvanced(ControlTestBase):
 
         # First restart to establish baseline
         self.log_step("Step 1: First exclusion and restart")
-        self.dmg.system_exclude(ranks=[test_rank], rank_hosts=None)
-        self.wait_for_rank_state(test_rank, "adminexcluded", timeout=10)
-        self.dmg.system_clear_exclude(ranks=[test_rank], rank_hosts=None)
-        self.wait_for_rank_state(test_rank, "joined", timeout=30)
+
+        restarted, final_state = self.exclude_rank_and_wait_restart(test_rank)
+
+        if not restarted:
+            self.fail("Rank %s did not automatically restart. Final state: %s"
+                      % (test_rank, final_state))
 
         first_restart_time = time.time()
 
-        # Second restart to measure delay
-        self.log_step("Step 2: Second exclusion to trigger deferred restart")
-        self.dmg.system_exclude(ranks=[test_rank], rank_hosts=None)
-        self.wait_for_rank_state(test_rank, "adminexcluded", timeout=10)
-        self.dmg.system_clear_exclude(ranks=[test_rank], rank_hosts=None)
-
-        # Wait for deferred restart
-        self.log_step("Step 3: Waiting for deferred restart (expected delay: %ss)",
+        # Second restart to measure delay and wait for deferred restart
+        self.log_step("Step 2: Wait for deferred restart (expected delay: %ss)",
                       expected_delay)
-        wait_timeout = expected_delay + 20  # Add buffer
 
-        if not self.wait_for_rank_state(test_rank, "joined", timeout=wait_timeout):
-            self.fail("Rank %s did not restart within expected time" % test_rank)
+        restarted, final_state = self.exclude_rank_and_wait_restart(test_rank,
+                                                                    timeout=35)
+
+        if not restarted:
+            self.fail("Rank %s did not automatically restart again. Final state: %s"
+                      % (test_rank, final_state))
 
         second_restart_time = time.time()
         actual_delay = second_restart_time - first_restart_time
 
-        self.log.info("Measured delay: %.1fs (expected: ~%ss)", actual_delay, expected_delay)
+        self.log.info("Measured delay: %.1fs (expected: ~%ss)", actual_delay,
+                      expected_delay)
 
         # Verify delay is within acceptable range (80% to 120% of expected)
         min_delay = expected_delay * 0.8
@@ -239,39 +187,3 @@ class EngineAutoRestartAdvanced(ControlTestBase):
         else:
             self.log.info("SUCCESS: Restart delay within expected range [%.1fs, %.1fs]",
                           min_delay, max_delay)
-
-#    def test_restart_after_clear_exclude(self):
-#        """Test interaction between auto-restart and manual clear-exclude.
-#
-#        Test Description:
-#            1. Exclude rank, wait for self-termination
-#            2. Clear exclusion before auto-restart triggers
-#            3. Verify rank rejoins successfully
-#
-#        :avocado: tags=all,daily_regression
-#        :avocado: tags=hw,medium
-#        :avocado: tags=dmg,control,engine_auto_restart
-#        :avocado: tags=EngineAutoRestartAdvanced,test_restart_after_clear_exclude
-#        """
-#        all_ranks = self.get_all_ranks()
-#        if len(all_ranks) < 2:
-#            self.skipTest("Test requires at least 2 ranks")
-#
-#        test_rank = self.random.choice(all_ranks)
-#
-#        self.log_step("Step 1: Excluding rank %s", test_rank)
-#        self.dmg.system_exclude(ranks=[test_rank], rank_hosts=None)
-#
-#        # Wait for self-termination
-#        if not self.wait_for_rank_state(test_rank, "adminexcluded", timeout=10):
-#            self.fail("Rank %s did not self-terminate" % test_rank)
-#
-#        # Clear exclusion before auto-restart
-#        self.log_step("Step 2: Clearing exclusion for rank %s", test_rank)
-#        self.dmg.system_clear_exclude(ranks=[test_rank], rank_hosts=None)
-#
-#        # Verify rank rejoins
-#        if not self.wait_for_rank_state(test_rank, "joined", timeout=30):
-#            self.fail("Rank %s did not rejoin after clear-exclude" % test_rank)
-#
-#        self.log.info("SUCCESS: Rank %s successfully rejoined after clear-exclude", test_rank)
